gridseed: Reduce traffic sending work to GridSeed devices

+++ gc3355.h
@@ -41,7 +41,7 @@ struct gc3355_orb_info
 #define gc3355_open(path)  serial_open(path, 115200, 1, true)
 #define gc3355_close(fd)  serial_close(fd)
 
-extern int gc3355_read(int fd, char *buf, size_t size);
+extern ssize_t gc3355_read(int fd, char *buf, size_t size);
 extern ssize_t gc3355_write(int fd, const void * const buf, const size_t size);
 
 extern void gc3355_init_usborb(int fd, int pll_freq, bool scrypt_only, bool detect_only);
